# internal/web/docs/openapi.yaml
openapi: 3.1.0
info:
  version: 0.0.1
  title: Leagueify API
  summary: Open Source Sporting League Platform.
  description: Host your own sports league using Leagueify, the all-in-one league hosting platform.
  license:
    name: MIT
    url: https://raw.githubusercontent.com/Leagueify/api/main/LICENSE
  contact:
    name: Leagueify Support
    url: https://github.com/Leagueify/api/issues
servers:
  - url: http://localhost/api
    description: Local Development Server - Docker
  - url: http://localhost:8888/api
    description: Local Development Server - Standalone Executable
paths:
  /accounts:
    post:
      tags:
      - Accounts
      summary: Create an account
      description: '
        Create an account for the Leagueify instance.
        Accounts created before the email service is configured do not require verification.
        **NOTE: Must be 18 or older to create an account.**
        '
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  description: User's first name
                  type: string
                  minLength: 3
                  maxLength: 32
                lastName:
                  description: User's last name
                  type: string
                  minLength: 3
                  MaxLength: 32
                email:
                  description: Email associated to the account
                  type: string
                password:
                  description: Raw user password for the account
                  type: string
                  minLength: 8
                dateOfBirth:
                  description: User's date of birth
                  type: string
                  example: "YYYY-MM-DD"
                phone:
                  description: User's phone number
                  type: string
                  example: "+12085550000"
                coach:
                  description: User's preference to coach
                  type: boolean
                  default: false
                volunteer:
                  description: User's preference to volunteer
                  type: boolean
                  default: false
              required:
                - firstName
                - lastName
                - email
                - password
                - dateOfBirth
                - phone
            examples:
              valid payload:
                summary: Valid account creation payload
                value: {
                  "firstName": "Leagueify",
                  "lastName": "User",
                  "email": "test@leagueify.org",
                  "password": "Test123!",
                  "dateOfBirth": "1990-08-31",
                  "phone": "+120855501234",
                  "coach": true,
                  "volunteer": false
                }
      responses:
        201:
          description: Account Created
          content:
            application/json:
              schema:
                $ref: "#/components/successful/schema"
              examples:
                accountCreated:
                  $ref: "#/components/successful/example"
        400:
          $ref: "#/components/errors/badRequest"
  /accounts/login:
    post:
      tags:
        - Accounts
      summary: Account login
      description: '
        Login to a created account for the Leagueify instance.
        This will return an API key that shall be used for API requests.
        '
      produces:
        - application/json
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  description: Email associated to the account
                  type: string
                password:
                  description: Raw user password for the account
                  type: string
              required:
                - email
                - password
            examples:
              valid payload:
                summary: Valid account login payload
                value: {
                  "email": "test@leagueify.org",
                  "password": "Test123!",
                }
      responses:
        200:
          description: Account Login Successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    description: Status text for account login outcome
                    type: string
                  apikey:
                    description: API Key used for API Authentication
                    type: string
              examples:
                accountLoginResponse:
                  summary: Account login successful
                  value: {
                    "status": "successful",
                    "apikey": "BHS4R8N34XQGCGEGWH2H2EDV77QR6TC59169WWPDAPX9F4JB1RETZJ7AEZEDVE1S"
                    }
        400:
          $ref: "#/components/errors/badRequest"
        401:
          $ref: "#/components/errors/unauthorized"
  /accounts/logout:
    post:
      tags:
        - Accounts
      summary: Account logout
      security:
        - apiKey: []
      description: '
        Logout of an account for the Leagueify instance.
        This will return an empty body.
        '
      produces:
        - application/json
      responses:
        200:
          description: Account Logout Successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    description: Status text for account logout
                    type: string
              examples:
                accountLogoutSuccessful:
                  summary: Account logout successful
                  value: {
                    "status": "successful",
                  }
        401:
          $ref: "#/components/errors/unauthorized"
  /accounts/{id}/verify:
    post:
      tags:
        - Accounts
      summary: Verify an account
      description: '
        Verify a created account for the Leagueify instance.
        Accounts created before the email service is configured do not require verification.
        '
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: ID of the account to verify
          required: true
          type: string
      responses:
        200:
          description: Account successfully verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    description: Status text for account verification outcome
                    type: string
                  apikey:
                    description: API Key used for API Authentication
                    type: string
              examples:
                accountCreationResponse:
                  summary: Account successfully verified
                  value: {
                    "status": "successful",
                    "apikey": "BHS4R8N34XQGCGEGWH2H2EDV77QR6TC59169WWPDAPX9F4JB1RETZJ7AEZEDVE1S"
                    }
        401:
          $ref: "#/components/errors/unauthorized"
...
components:
  errors:
    badRequest:
      description: Error with request body
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                description: Error message detailing the request issue
                type: string
          examples:
            default:
              summary: Response containing error message
              value: {
                "detail": "< details about the error >",
                "status": "bad request"
                }
    unauthorized:
      description: Authorization Issue
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                description: Error message detailing the request issue
                type: string
          examples:
            default:
              summary: Response containing unauthorized message
              value: {
                "status": "unauthorized"
                }
...
  successful:
    schema:
      type: object
      properties:
        status:
          description: Successful Request Response
          type: string
    example:
      summary: Successful Request Response
      value: {
          "status": "successful"
        }
  securitySchemes:
    apiKey:
      type: apiKey
      name: apiKey
      in: header